# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "Complete API for mechanic shop management with inventory tracking, JWT authentication, rate limiting, and caching"
  version: "1.0.0"
  contact:
    name: "Mechanic Shop API Support"
    email: "support@mechanicshop.com"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      JWT Authorization header using the Bearer scheme.
      
      **Format:** `Bearer {jwt_token}`
      
      **Example:** `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
      
      **How to obtain token:**
      1. Make a POST request to `/login` with valid credentials
      2. Copy the `access_token` from the response
      3. Add it to the Authorization header as `Bearer {access_token}`
      
      **Token contains:**
      - User identity and role information
      - Expiration time (typically 24 hours)
      - Permissions and access level
      
      **Usage:**
      - Include this header in requests to protected endpoints
      - Token will be validated on each request
      - Expired tokens will return 401 Unauthorized
  
  JWTAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      JSON Web Token (JWT) authentication for API access.
      
      **Authentication Flow:**
      1. **Login:** POST credentials to `/login` endpoint
      2. **Token:** Receive JWT access token in response
      3. **Header:** Include token in Authorization header
      4. **Access:** Make authenticated requests to protected resources
      
      **Token Security:**
...
paths:
...
  /members:
    post:
      tags:
        - "Members"
      summary: "Create new member"
      description: |
        Create a new member in the system with comprehensive profile information.
        
        **Purpose:**
        - Register new members with complete profile data
        - Establish member identity and contact information
        - Set up membership preferences and settings
        - Create initial account with secure credentials
        
        **Features:**
        - Input validation for all required fields
        - Email uniqueness verification
        - Password security requirements
        - Membership type classification
        - Contact information management
        
        **Process:**
        1. Validate input payload structure
        2. Check email availability
        3. Encrypt password securely
        4. Create member record in database
        5. Return member details with unique ID
        
        **Rate Limiting:** 3 requests per minute per IP
      operationId: "createMember"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: memberData
          description: "Complete member information for account creation"
          required: true
          schema:
            $ref: "#/definitions/CreateMemberPayload"
      security:
        - Bearer: []
        - JWTAuth: []
      responses:
        201:
          description: "Member created successfully"
          schema:
            $ref: "#/definitions/CreateMemberResponse"
          examples:
            application/json:
              message: "Member created successfully"
              status: "success"
              member:
                id: 123
                first_name: "John"
                last_name: "Doe"
                email: "john.doe@example.com"
                phone_number: "555-0123"
                membership_type: "premium"
                join_date: "2024-12-24"
                is_active: true
              timestamp: "2024-12-24T10:30:00Z"
        400:
          description: "Bad request - Invalid input data"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Validation failed"
              message: "Invalid email format provided"
              timestamp: "2024-12-24T10:30:00Z"
        409:
          description: "Conflict - Email already exists"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Duplicate entry"
              message: "Member with this email already exists"
              timestamp: "2024-12-24T10:30:00Z"
        422:
          description: "Unprocessable entity - Missing required fields"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Missing required fields"
              message: "First name, last name, and email are required"
              timestamp: "2024-12-24T10:30:00Z"
        429:
          description: "Too many requests - Rate limit exceeded"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Rate limit exceeded"
              message: "Maximum 3 member creation requests per minute"
              timestamp: "2024-12-24T10:30:00Z"
    get:
      tags:
        - "Members"
      summary: "Get all members"
      description: "Retrieve a list of all members in the system"
      security:
        - Bearer: []
        - JWTAuth: []
      responses:
        200:
          description: "Successfully retrieved all members"
          schema:
            $ref: "#/definitions/AllMembers"
  /members/{id}:
    get:
      tags:
        - "Members"
      summary: "Get member by ID"
      description: |
        Retrieve detailed information about a specific member by their unique ID.
        
        **Purpose:**
        - Get complete member profile
        - View membership details and status
        - Access member history and preferences
        
        **Security:** Basic member info accessible, sensitive data requires authentication
      operationId: "getMemberById"
      produces:
        - "application/json"
      security:
        - Bearer: []
        - JWTAuth: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: "Unique member ID"
          minimum: 1
      responses:
        200:
          description: "Member retrieved successfully"
          schema:
            $ref: "#/definitions/Member"
          examples:
            application/json:
              id: 123
              first_name: "John"
              last_name: "Doe"
              email: "john.doe@example.com"
              membership_type: "premium"
              join_date: "2024-12-24"
              is_active: true
              member_number: "MEM-2024-000123"
        404:
          description: "Member not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Member not found"
              message: "No member found with ID 123"
              timestamp: "2024-12-24T10:30:00Z"
    put:
      tags:
        - "Members"
      summary: "Update member"
      description: |
        Update member information and preferences with comprehensive validation.
        
        **Purpose:**
        - Update member profile information selectively
        - Change membership type and billing preferences
        - Modify contact information and address
        - Update account preferences and settings
        - Maintain audit trail of changes
        
        **Features:**
        - Partial updates (only send fields to change)
        - Input validation for all fields
        - Email uniqueness verification (if email updated)
        - Password encryption (if password updated)
        - Membership type validation
        - Preference management
        
        **Security:** 
        - Requires valid JWT authentication
        - Members can update own profile
        - Admin/Manager can update any member
        - Sensitive operations require elevated permissions
        
        **Process:**
        1. Validate JWT token and member permissions
        2. Verify member exists and is accessible
        3. Validate input payload structure
        4. Check email availability (if email changed)
        5. Encrypt new password securely (if provided)
        6. Update member record with changes
        7. Return updated member details
        
        **Rate Limiting:** 5 updates per minute per member
        **Audit:** All updates are logged for security
      operationId: "updateMember"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: "Unique member ID to update"
          minimum: 1
        - in: body
          name: memberUpdate
          description: |
            Member information to update. Only include fields you want to change.
            
            **Partial Updates Supported:**
            - Send only the fields you want to update
            - Omitted fields will remain unchanged
            - Empty/null values will be ignored
            
            **Field Validation:**
            - Email must be unique across all members
            - Password must meet security requirements (8+ chars)
            - Membership type must be valid enum value
            - Phone number format validation applied
            
            **Security Notes:**
            - New passwords are automatically encrypted
            - Sensitive fields require elevated permissions
            - All changes are audit logged
          required: true
          schema:
            $ref: "#/definitions/UpdateMemberPayload"
      responses:
        200:
          description: "Member updated successfully"
          schema:
            $ref: "#/definitions/UpdateMemberResponse"
          examples:
            application/json:
              message: "Member updated successfully"
              status: "success"
              member:
                id: 123
                first_name: "Jane"
                last_name: "Smith"
                email: "jane.smith@example.com"
                phone_number: "555-9876"
                membership_type: "gold"
                join_date: "2024-12-24"
                last_updated: "2024-12-25T14:30:00Z"
                is_active: true
                member_number: "MEM-2024-000123"
              timestamp: "2024-12-25T14:30:00Z"
        400:
          description: "Bad request - Invalid update data"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Validation failed"
              message: "Email format is invalid"
              timestamp: "2024-12-25T14:30:00Z"
        401:
          description: "Unauthorized - Authentication required"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Authentication required"
              message: "Valid JWT token required to update member"
              timestamp: "2024-12-25T14:30:00Z"
        403:
          description: "Forbidden - Cannot update this member"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Access denied"
              message: "You do not have permission to update this member"
              timestamp: "2024-12-25T14:30:00Z"
        404:
          description: "Member not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        429:
          description: "Too many requests - Rate limit exceeded"
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - "Members"
      summary: "Delete member"
      description: |
        Permanently delete a member account and all associated data.
        
        **Purpose:**
        - Remove member from system
        - Delete all member-related data
        - Handle account closure requests
        
        **Security:** Requires admin authentication
        **Warning:** This action is irreversible
      operationId: "deleteMember"
      produces:
        - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: "Unique member ID to delete"
          minimum: 1
      responses:
        200:
          description: "Member deleted successfully"
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Member deleted successfully"
              deleted_member_id:
                type: integer
                example: 123
              timestamp:
                type: string
                format: date-time
                example: "2024-12-24T10:30:00Z"
        401:
          description: "Unauthorized - Admin authentication required"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Forbidden - Insufficient permissions"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Member not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        description: "Error category or type"
        example: "Authentication failed"
      message:
        type: string
        description: "Detailed error message"
        example: "Invalid email or password"
      timestamp:
        type: string
        format: date-time
        description: "When the error occurred"
        example: "2024-12-24T10:30:00Z"
    example:
      error: "Authentication failed"
      message: "Invalid email or password"
      timestamp: "2024-12-24T10:30:00Z"

  # Member Management Definitions
  CreateMemberPayload:
    type: object
    required:
      - first_name
      - last_name
      - email
      - password
      - phone_number
      - membership_type
    properties:
      first_name:
        type: string
        description: "Member's first name"
        example: "John"
        minLength: 2
        maxLength: 50
      last_name:
        type: string
        description: "Member's last name"
        example: "Doe"
        minLength: 2
        maxLength: 50
      email:
        type: string
        format: email
        description: "Member's email address (must be unique)"
        example: "john.doe@example.com"
        minLength: 5
        maxLength: 255
      password:
        type: string
        description: "Secure password for member account (minimum 8 characters)"
        example: "SecurePass123!"
        minLength: 8
        maxLength: 255
      phone_number:
        type: string
        description: "Member's contact phone number"
        example: "555-0123"
        minLength: 10
        maxLength: 20
      membership_type:
        type: string
        description: "Type of membership subscription"
        example: "premium"
        enum: ["basic", "premium", "gold", "platinum"]
      date_of_birth:
        type: string
        format: date
        description: "Member's date of birth (optional)"
        example: "1990-05-15"
      address:
        type: object
        description: "Member's address information (optional)"
        properties:
          street:
            type: string
            example: "123 Main St"
          city:
            type: string
            example: "Anytown"
          state:
            type: string
            example: "CA"
          zip_code:
            type: string
            example: "12345"
          country:
            type: string
            example: "USA"
      preferences:
        type: object
        description: "Member preferences and settings (optional)"
        properties:
          newsletter_subscription:
            type: boolean
            example: true
          notifications_enabled:
            type: boolean
            example: true
          preferred_language:
            type: string
            example: "en"
    example:
      first_name: "John"
      last_name: "Doe"
      email: "john.doe@example.com"
      password: "SecurePass123!"
      phone_number: "555-0123"
      membership_type: "premium"
      date_of_birth: "1990-05-15"
      address:
        street: "123 Main St"
        city: "Anytown"
        state: "CA"
        zip_code: "12345"
        country: "USA"
      preferences:
        newsletter_subscription: true
        notifications_enabled: true
        preferred_language: "en"
  CreateMemberResponse:
    type: object
    required:
      - message
      - status
      - member
      - timestamp
    properties:
      message:
        type: string
        description: "Success message confirming member creation"
        example: "Member created successfully"
      status:
        type: string
        description: "Operation status indicator"
        example: "success"
        enum: ["success", "error"]
      member:
        type: object
        description: "Created member information (excluding sensitive data)"
        required:
          - id
          - first_name
          - last_name
          - email
          - membership_type
          - join_date
          - is_active
        properties:
          id:
            type: integer
            description: "Unique member identifier"
            example: 123
          first_name:
            type: string
            description: "Member's first name"
            example: "John"
          last_name:
            type: string
            description: "Member's last name"
            example: "Doe"
          email:
            type: string
            format: email
            description: "Member's email address"
            example: "john.doe@example.com"
          phone_number:
            type: string
            description: "Member's phone number"
            example: "555-0123"
          membership_type:
            type: string
            description: "Type of membership"
            example: "premium"
          join_date:
            type: string
            format: date
            description: "Date when member joined"
            example: "2024-12-24"
          is_active:
            type: boolean
            description: "Whether the member account is active"
            example: true
          member_number:
            type: string
            description: "Unique member number for identification"
            example: "MEM-2024-000123"
      timestamp:
        type: string
        format: date-time
        description: "When the member was created"
        example: "2024-12-24T10:30:00Z"
    example:
      message: "Member created successfully"
      status: "success"
      member:
        id: 123
        first_name: "John"
        last_name: "Doe"
        email: "john.doe@example.com"
        phone_number: "555-0123"
        membership_type: "premium"
        join_date: "2024-12-24"
        is_active: true
        member_number: "MEM-2024-000123"
      timestamp: "2024-12-24T10:30:00Z"
  UpdateMemberPayload:
    type: object
    description: "Payload for updating an existing member - all fields are optional for partial updates"
    properties:
      first_name:
        type: string
        description: "Member's first name"
        minLength: 2
        maxLength: 50
        example: "John"
      last_name:
        type: string
        description: "Member's last name"
        minLength: 2
        maxLength: 50
        example: "Doe"
      email:
        type: string
        format: email
        description: "Member's email address (must be unique)"
        example: "john.doe@example.com"
      phone_number:
        type: string
        description: "Member's phone number"
        minLength: 10
        maxLength: 20
        example: "555-0123"
      membership_type:
        type: string
        enum: ["basic", "premium", "gold", "platinum"]
        description: "Type of membership"
        example: "premium"
      date_of_birth:
        type: string
        format: date
        description: "Member's date of birth"
        example: "1990-01-15"
      is_active:
        type: boolean
        description: "Whether the member is active"
        example: true
      role:
        type: string
        enum: ["member", "admin", "manager"]
        description: "Member's role in the system"
        example: "member"
    example:
      first_name: "John"
      last_name: "Doe"
      email: "john.doe@example.com"
      phone_number: "555-0123"
      membership_type: "premium"
      is_active: true
      role: "member"
  UpdateMemberResponse:
    type: object
    required: ["message", "status", "member", "timestamp"]
    description: "Response after successfully updating a member"
    properties:
      message:
        type: string
        description: "Success message confirming the update"
        example: "Member updated successfully"
      status:
        type: string
        description: "Operation status indicator"
        enum: ["success", "error"]
        example: "success"
      member:
        $ref: "#/definitions/MemberResponse"
        description: "Updated member object (excluding sensitive data)"
      timestamp:
        type: string
        format: date-time
        description: "When the update occurred"
        example: "2024-12-24T10:30:00Z"
      updated_fields:
        type: array
        items:
          type: string
        description: "List of fields that were updated"
        example: ["first_name", "email", "membership_type"]
    example:
      message: "Member updated successfully"
      status: "success"
      member:
        id: 123
        first_name: "John"
        last_name: "Doe"
        email: "john.doe@example.com"
        phone_number: "555-0123"
        membership_type: "premium"
        role: "member"
        is_active: true
        join_date: "2024-01-15"
        updated_at: "2024-12-24T10:30:00Z"
      timestamp: "2024-12-24T10:30:00Z"
      updated_fields: ["first_name", "email", "membership_type"]